+++ packages/core/src/plugins/html-deserializer/utils/collapse-white-space/collapseWhiteSpace.ts
@@ -1,0 +1,17 @@
+import { collapseWhiteSpaceElement } from './collapseWhiteSpaceElement';
+import { CollapseWhiteSpaceState } from './types';
+
+// Entrypoint
+export const collapseWhiteSpace = (element: HTMLElement) => {
+  const clonedElement = element.cloneNode(true) as HTMLElement;
+
+  // Mutable state object
+  const state: CollapseWhiteSpaceState = {
+    inlineFormattingContext: null,
+    whiteSpaceRule: 'normal',
+  };
+
+  collapseWhiteSpaceElement(clonedElement, state);
+
+  return clonedElement;
+};
