+++ packages/core/src/plugins/html-deserializer/utils/collapse-white-space/collapseWhiteSpaceNode.ts
@@ -1,0 +1,23 @@
+import { isHtmlElement } from '../isHtmlElement';
+import { isHtmlText } from '../isHtmlText';
+import { collapseWhiteSpaceChildren } from './collapseWhiteSpaceChildren';
+import { collapseWhiteSpaceElement } from './collapseWhiteSpaceElement';
+import { collapseWhiteSpaceText } from './collapseWhiteSpaceText';
+import { CollapseWhiteSpaceState } from './types';
+
+export const collapseWhiteSpaceNode = (
+  node: Node,
+  state: CollapseWhiteSpaceState
+) => {
+  if (isHtmlElement(node)) {
+    collapseWhiteSpaceElement(node as HTMLElement, state);
+    return;
+  }
+
+  if (isHtmlText(node)) {
+    collapseWhiteSpaceText(node as Text, state);
+    return;
+  }
+
+  collapseWhiteSpaceChildren(node, state);
+};
