+++ packages/core/src/plugins/html-deserializer/utils/collapse-white-space/index.ts
@@ -1,0 +1,17 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from './collapseString';
+export * from './collapseWhiteSpace';
+export * from './collapseWhiteSpaceChildren';
+export * from './collapseWhiteSpaceElement';
+export * from './collapseWhiteSpaceNode';
+export * from './collapseWhiteSpaceText';
+export * from './inferWhiteSpaceRule';
+export * from './inlineTagNames';
+export * from './isHtmlBlockElement';
+export * from './isHtmlInlineElement';
+export * from './isLastNonEmptyTextOfInlineFormattingBlock';
+export * from './stateTransforms';
+export * from './types';
