+++ packages/core/src/plugins/html-deserializer/utils/collapse-white-space/isHtmlBlockElement.ts
@@ -1,0 +1,8 @@
+import { isHtmlElement } from '../isHtmlElement';
+import { isHtmlInlineElement } from './isHtmlInlineElement';
+
+export const isHtmlBlockElement = (node: Node): boolean => {
+  if (!isHtmlElement(node)) return false;
+  const element = node as HTMLElement;
+  return !isHtmlInlineElement(element);
+};
