+++ packages/core/src/plugins/html-deserializer/utils/collapse-white-space/stateTransforms.ts
@@ -1,0 +1,18 @@
+import { CollapseWhiteSpaceState } from './types';
+
+export const upsertInlineFormattingContext = (
+  state: CollapseWhiteSpaceState
+) => {
+  if (state.inlineFormattingContext) {
+    state.inlineFormattingContext.atStart = false;
+  } else {
+    state.inlineFormattingContext = {
+      atStart: true,
+      lastHasTrailingWhiteSpace: false,
+    };
+  }
+};
+
+export const endInlineFormattingContext = (state: CollapseWhiteSpaceState) => {
+  state.inlineFormattingContext = null;
+};
