+++ packages/core/src/plugins/html-deserializer/utils/index.ts
@@ -8,7 +8,6 @@ export * from './cleanHtmlEmptyElements';
 export * from './cleanHtmlFontElements';
 export * from './cleanHtmlLinkElements';
 export * from './cleanHtmlTextNodes';
-export * from './collapseWhiteSpace';
 export * from './copyBlockMarksToSpanChild';
 export * from './deserializeHtml';
 export * from './deserializeHtmlElement';
@@ -44,3 +43,4 @@ export * from './traverseHtmlElements';
 export * from './traverseHtmlNode';
 export * from './traverseHtmlTexts';
 export * from './unwrapHtmlElement';
+export * from './collapse-white-space/index';
